+++ src/app/account/login/page.tsx
@@ -1,44 +1,10 @@
 import { Metadata } from 'next';
-import Link from 'next/link';
-import { LoginForm } from '@/app/account/login/login-form';
+import { Login } from '@/app/account/login/login';
 
 export const metadata: Metadata = {
     title: 'Login',
 };
 
 export default async function LoginPage() {
-    return (
-        <main className="mx-auto max-w-screen-sm space-y-6 p-6">
-            <h1 className="scroll-m-20 text-4xl font-extrabold tracking-tight lg:text-5xl">
-                Login
-            </h1>
-
-            <p className="leading-7">
-                Enter your email address below to login to your account.
-            </p>
-
-            <LoginForm />
-
-            <div className="space-y-2">
-                <p className="text-muted-foreground text-sm">
-                    Not already a member?{' '}
-                    <Link
-                        href="/account/register"
-                        className="underline underline-offset-4"
-                    >
-                        Register now
-                    </Link>
-                </p>
-                <p className="text-muted-foreground text-sm">
-                    Forgotten your password?{' '}
-                    <Link
-                        href="/account/forgot-password"
-                        className="underline underline-offset-4"
-                    >
-                        Request reset
-                    </Link>
-                </p>
-            </div>
-        </main>
-    );
+    return <Login />;
 }
