+++ src/app/profile/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 import { getProfileAction } from '@/app/profile/actions/get-profile-action';
-import Profile from '@/app/profile/profile';
+import { Profile } from '@/app/profile/profile';
 import {
     isServerActionSuccess,
     ServerActionError,
